Update system_install_ubuntu_1804.rst
The guide for installing Elasticsearch 7.5 will actually install 7.8.x which is not compatible with Akeneo 4.x one issue is that bulk action is not working. To install 7.5.x you need to manually install the package from elastic. Even better if Akeneo was updated to support the latest version of ES.